refactor(app): seat the local actor, region_gap and base64_len with their callers
Three pure moves from the review's second bug pass. local_actor and
hook_agent_actor follow the in-process door into gateway/local_action.rs,
whose doc already claimed the trusted actor; nothing widens, the child reads
the root's private constants directly. region_gap returns to evidence.rs
beside screenshot_gap, private, since both its callers are in the root;
base64_len moves to evidence/image.rs, the one file that calls it.
